% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richtext_helpers.R
\name{rich_str_wrap}
\alias{rich_str_wrap}
\title{str_wrap but for richtext}
\usage{
rich_str_wrap(string, width)
}
\arguments{
\item{string}{(character) the string to be wrapped}

\item{width}{(numeric) the width to be wrapped}
}
\value{
a string
}
\description{
This function serves the same purpose as \code{stringr::str_wrap} but
replaces the \verb{\\n} breaks with \verb{<br>} breaks for use in a richtext format
}
\examples{
rich_str_wrap(c("This is a sentence",
          "This is another sentence"),
          5)
}
